% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_lib.R
\name{del_country_name}
\alias{del_country_name}
\title{Del_country_name}
\usage{
del_country_name(posnames)
}
\arguments{
\item{posnames}{Character vector of possible countries names.}
}
\value{
Unit character vector "Success" if deleting was correct, or unit vector "Invalid arguments" if type of argument isn`t correct (for example, zero length names in argument posnames).
}
\description{
Delete countries names from database
}
\examples{
del_country_name("Russia")
vect <- c("ru", "en", "usa") # posnames
del_country_name(vect)
}
