% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_lib.R
\name{match_country_name}
\alias{match_country_name}
\title{Match_country_name}
\usage{
match_country_name(posnames, realnames, priorities = "2")
}
\arguments{
\item{posnames}{Character vector of possible countries names.}

\item{realnames}{Character vector of official(common) countries names.}

\item{priorities}{Character vector of priorities. Takes values "1" or "2". Priorities should have same length with posnames, otherwise will be initialized priorities vector with length of posnames and value "2".}
}
\value{
Unit character vector "Success" if matching was correct, or unit vector "Invalid arguments" if types of arguments aren`t correct (for example, zero length names in argument posnames or diffrent length of posnames and realnames).
}
\description{
Add countries names to database
}
\examples{
match_country_name("Rassiya", "Russia")
vect_1 <- c("ru", "en", "usa") # posnames
vetc_2 <- c("Russia", "United Kingdom", "United States") #realnames
match_country_name(vect_1, vect_2)
vect_3 <- c("1", "2", "1") # priorities
match_country_name(vect_1, vect_2, vect_3)
}
