% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_lib.R
\name{normalize_country_name}
\alias{normalize_country_name}
\title{Normalize_country_name}
\usage{
normalize_country_name(posnames, dif_acc = 0.2)
}
\arguments{
\item{posnames}{Character vector of countries names to normalization.}

\item{dif_acc}{Float accuracy parametr of string matching. Must be from 0.0 to 1.0}
}
\value{
Character vector of normalized names. Contains common names if normalizing was successful and "None" if not or unit vector "Invalid arguments" if types of arguments aren`t correct (for example, zero length names in argument posnames or incorrect type of dif_acc).
}
\description{
Countries names normalization (find the correct)
}
\examples{
normalize_country_name("Russia")
vect <- c("ru", "en", "usa") # posnames
normalize_country_name(vect)
normalize_country_name("United States", 0.3)
}
